# openapi/cookingchallenge-api.yaml
openapi: 3.0.3
info:
  title: CookingChallenge API
  description: >
    API contract for the CookingChallenge organizer/guest cooking-challenge app.
    Shaped around the screens in the UI mockup (see
    docs/cookingChallenge/design-reference.md) rather than a 1:1 dump of the domain
    model: every computed value (submission progress, category winners, rivalry
    summaries) is pre-shaped by the server so the frontend only holds UI state.
  version: "1.0.0"
servers:
  - url: /
...
paths:
...
  /api/v1/auth/access-link-login:
    post:
      tags: [Auth]
      operationId: accessLinkLogin
      summary: Exchange a personalized access-link token for a JWT
      description: >
        Guests authenticate via a personalized access link emailed to them (see
        `sendInvitations`), but from that point on use the same JWT bearer mechanism as
        organizers - this endpoint is the one-time exchange step. The link token itself
        remains valid (reusable until its own expiry) and can be exchanged again to obtain a
        fresh JWT.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccessLinkLoginRequest'
      responses:
        '200':
          description: Authenticated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/auth/password-reset:
    post:
      tags: [Auth]
      operationId: redeemPasswordReset
      summary: Set a new password using a password-reset token
      description: >
        Redeems the token from a reset mail issued by `triggerPasswordReset`. Tokens are
        single-use and expire after 2 hours; an unknown, expired or already-used token is
        rejected with the same `INVALID_OR_EXPIRED_LINK` error, deliberately
        indistinguishable. Confirming the new password twice is a UI concern and is not part
        of this contract.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetRedeemRequest'
      responses:
        '204':
          description: Password changed. The caller logs in with it via `login`.
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/accounts:
    get:
      tags: [Accounts]
      operationId: listAccounts
      summary: List accounts
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        '200':
          description: Paginated list of accounts.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [Accounts]
      operationId: createAccount
      summary: Create an account
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAccountRequest'
      responses:
        '201':
          description: Account created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/accounts/{accountId}:
    get:
      tags: [Accounts]
      operationId: getAccount
      summary: Get a single account, for the edit dialog to load fresh data on open
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/AccountIdPathParam'
      responses:
        '200':
          description: Account detail. Selectable roles come from GET /api/v1/config, not this response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountDetailResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Accounts]
      operationId: updateAccount
      summary: Edit an account's name/email/roles
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/AccountIdPathParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAccountRequest'
      responses:
        '200':
          description: Account updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/accounts/{accountId}/password-reset:
    post:
      tags: [Accounts]
      operationId: triggerPasswordReset
      summary: Email an account a link to set a new password (admin only)
      description: >
        Only accounts that hold a password (organizers and admins) can be reset; guest
        accounts log in through access links and are rejected with 409
        `PASSWORD_RESET_NOT_ELIGIBLE`. Issuing a new reset invalidates any earlier,
        unused reset link for the same account. Delivery is asynchronous and best-effort,
        hence 202 with no body.
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/AccountIdPathParam'
      responses:
        '202':
          description: Reset link issued; the mail is on its way.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: The account has no password to reset (a guest account).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Login token — issued by POST /api/v1/auth/login (organizer/admin, password) or
        POST /api/v1/auth/access-link-login (guest, exchanging a personalized access-link
        token). Every authenticated endpoint uses this same mechanism.
  parameters:
    PageParam:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
    SizeParam:
      name: size
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    AccountIdPathParam:
      name: accountId
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/Tsid'
...
  responses:
    ValidationError:
      description: Request failed validation.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Missing, invalid, or expired credentials.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Authenticated but not permitted to perform this action.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource does not exist.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Request conflicts with current state.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    Tsid:
      type: string
      description: Aggregate id, Crockford Base32-encoded TSID.
      example: "0K3X8Y4Z9QJ7M"
    ApiMeta:
      type: object
      required: [requestId, timestamp]
      properties:
        requestId:
          type: string
        timestamp:
          type: string
          format: date-time
    Pagination:
      type: object
      required: [page, size, totalElements, totalPages, first, last]
      properties:
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        first:
          type: boolean
        last:
          type: boolean
    ErrorDetail:
      type: object
      required: [field, message]
      properties:
        field:
          type: string
        message:
          type: string
    ErrorBody:
      type: object
      required: [code, message, details, requestId, timestamp]
      properties:
        code:
          type: string
          example: NOT_FOUND
        message:
          type: string
        details:
          type: array
          items:
            $ref: '#/components/schemas/ErrorDetail'
        requestId:
          type: string
        timestamp:
          type: string
          format: date-time
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/ErrorBody'

    # --- Domain vocabulary ---
...
    SystemRole:
      type: string
      enum: [ADMIN, ORGANIZER, USER]
...
    AccessLinkLoginRequest:
      type: object
      required: [token]
      properties:
        token:
          type: string
    PasswordResetRedeemRequest:
      type: object
      required: [token, newPassword]
      properties:
        token:
          type: string
        newPassword:
          type: string
          minLength: 8
    AuthToken:
      type: object
      required: [accessToken, expiresAt]
      properties:
        accessToken:
          type: string
        expiresAt:
          type: string
          format: date-time
    AuthTokenResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          $ref: '#/components/schemas/AuthToken'
        meta:
          $ref: '#/components/schemas/ApiMeta'

    # --- Accounts ---
    Account:
      type: object
      required: [id, email, firstName, lastName, name, roles]
      properties:
        id:
          $ref: '#/components/schemas/Tsid'
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        name:
          type: string
          readOnly: true
          description: Derived as "{firstName} {lastName}" — not settable, see CreateAccountRequest/UpdateAccountRequest.
        roles:
          type: array
          items:
            $ref: '#/components/schemas/SystemRole'
    CreateAccountRequest:
      type: object
      required: [email, firstName, lastName]
      properties:
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        roles:
          type: array
          items:
            $ref: '#/components/schemas/SystemRole'
          description: Defaults to [USER] if omitted.
        password:
          type: string
          nullable: true
          description: Required for ORGANIZER/ADMIN roles; ignored for USER-only accounts.
    UpdateAccountRequest:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        roles:
          type: array
          items:
            $ref: '#/components/schemas/SystemRole'
          description: Replaces the account's full role set; must not end up empty.
    AccountResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          $ref: '#/components/schemas/Account'
        meta:
          $ref: '#/components/schemas/ApiMeta'
    AccountDetailResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          $ref: '#/components/schemas/Account'
        meta:
          $ref: '#/components/schemas/ApiMeta'
    AccountListResponse:
      type: object
      required: [data, pagination, meta]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Account'
        pagination:
          $ref: '#/components/schemas/Pagination'
        meta:
          $ref: '#/components/schemas/ApiMeta'

    # --- Challenges (organizer-facing) ---
...
security:
  - bearerAuth: []